Reorganize structure, delete unneeded code
