+++ util/analyze/__init__.py
@@ -1,8 +1,3 @@
-from .analyzer import Analyzer
-from .logs import Logs, Benchmark, Block
-
-from ._main import main, basemain, parse_logs, parse_args
-
-import sys
-
-__all__ = ['logs', 'Analyzer', 'Logs', 'Benchmark', 'Block']
+from ._types import Logs, Benchmark, Block
+from ._main import parse_logs, parse_args
+from .imports import import_cpu2006, import_plaidml, import_shoc, import_utils
