+++ util/analyze/_main.py
@@ -1,15 +1,9 @@
-import sys
 import pickle
 import argparse
 import json
 import fnmatch
 
-from .analyzer import Analyzer
-
-from . import import_cpu2006
-from . import import_plaidml
-from . import import_shoc
-from . import import_utils
+from .imports import *
 
 
 def __load_file(file):
@@ -26,6 +20,7 @@ def __load_filepath(filepath):
 
 def parse_logs(filepath: str):
     FILE_PARSERS = {
+        'pickle': __load_filepath,
         'spec': import_cpu2006.parse,
         'plaidml': import_plaidml.parse,
         'shoc': import_shoc.parse,
@@ -46,7 +41,8 @@ def parse_logs(filepath: str):
 
     def blk_filter_f(blk):
         return all(
-            event in blk and all(log_matches(log, matcher) for log in blk[event])
+            event in blk and all(log_matches(log, matcher)
+                                 for log in blk[event])
             for event, matcher in blk_filter.items()
         )
 
@@ -62,7 +58,7 @@ def parse_args(parser: argparse.ArgumentParser, args=None):
         parser.add_argument(
             '--benchsuite',
             required=real,
-            choices=('spec', 'plaidml', 'shoc'),
+            choices=('spec', 'plaidml', 'shoc', 'pickle'),
             help='Select the benchmark suite which the input satisfies.',
         )
         parser.add_argument(
@@ -82,68 +78,3 @@ def parse_args(parser: argparse.ArgumentParser, args=None):
     result = parser.parse_args(args)
 
     return result
-
-
-def basemain(*, positional, options, description, action, manual_options={}):
-    parser = argparse.ArgumentParser(description=description)
-
-    for name, help in positional:
-        parser.add_argument(name, help=help)
-
-    for name, help in options.items():
-        if name in manual_options:
-            parser.add_argument(
-                '--' + name, default=manual_options[name], help=help)
-        else:
-            parser.add_argument('--' + name, help=help)
-
-    parser.add_argument(
-        '--benchsuite',
-        default=manual_options.get('benchsuite', None),
-        choices=('spec', 'plaidml', 'shoc'),
-        help='Select the benchmark suite which the input satisfies. Valid options: spec',
-    )
-    parser.add_argument(
-        '-o', '--output',
-        default=manual_options.get('output', None),
-        help='Where to output the report',
-    )
-
-    args = parser.parse_args()
-    option_values = {name: getattr(args, name) for name in options}
-    pos = [getattr(args, name) for name, help in positional]
-
-    FILE_PARSERS = {
-        None: __load_filepath,
-        'spec': import_cpu2006.parse,
-        'plaidml': import_plaidml.parse,
-        'shoc': import_shoc.parse,
-    }
-    parser = FILE_PARSERS[args.benchsuite]
-
-    pos_data = [parser(f) for f in pos]
-
-    if args.output is None:
-        outfile = sys.stdout
-    else:
-        outfile = open(args.output, 'w')
-
-    try:
-        action(outfile, pos_data, option_values)
-    finally:
-        outfile.close()
-
-
-def main(analyzer_cls, **manual_options):
-    def main_action(outfile, pos_data, option_values):
-        analyzer = analyzer_cls(**option_values)
-        analyzer.run(pos_data)
-        analyzer.print_report(outfile)
-
-    basemain(
-        positional=analyzer_cls.POSITIONAL,
-        options=analyzer_cls.OPTIONS,
-        description=analyzer_cls.__doc__,
-        action=main_action,
-        manual_options=manual_options,
-    )
