+++ util/analyze/_types.py
